requirements.txt
streamlitprimer3-pypandasplotlybiopython